change param type when using invoke

+++ src/main/java/methods/PanamaPowerCalculator.java
@@ -21,7 +21,7 @@ public class PanamaPowerCalculator {
         }
     }
 
-    public static double calculateInvoke(double base, double exponent) {
+    public static double calculateInvoke(int base, int exponent) {
         try {
             return (double) METHOD_HANDLE.invoke(base, exponent);
         } catch (Throwable t) {
